Program/qg_app_backend/requirements.txt
asgiref==3.6.0
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
filelock==3.9.0
huggingface-hub==0.13.1
idna==3.4
iniconfig==2.0.0
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==23.0
Pillow==9.4.0
pluggy==1.0.0
pytest==7.2.2
pytest-django==4.5.2
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tokenizers==0.13.2
torch==1.13.1
tqdm==4.42.1
transformers==4.26.1
typing_extensions==4.5.0
urllib3==1.26.14
